refactor(core): extract shared tree-sitter parser from code-similarity
The AstParser trait, TreeSitterParser, and ParseError lived inside the
code_similarity module but aren't specific to it. Moved them to a
top-level parser module in scute-core so upcoming checks can reuse
them without duplicating parser setup.

Part of #31
